step17の作業ファイルの内容を取り込み
